refactor(commands): extract /search command into search.rs
